Changed way the classes were called, made it more Object oriented.
Classes are now being initialized in the __init__.py file on the API package. they are then provided to the handlers that need them.

Handlers also have been given a constructor and the DAO class is being initialized in it, instead of having to create the object every time the call was made.

tested it on postman, works so far